requirements.txt
Flask==3.0.3
PyYAML==6.0
pycdlib==1.14.0
pefile==2023.2.7
gunicorn==21.2.0
